requirements.txt
livekit>=1.0
livekit-agents[openai,deepgram,cartesia,silero,turn_detector]~=1.0rc
python-dotenv~=1.0
